% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerationProbability.R
\name{accelerationProbability}
\alias{accelerationProbability}
\title{accelerationProbability}
\usage{
accelerationProbability(lower, upper, params)
}
\arguments{
\item{lower}{The lower bound of the range.}

\item{upper}{The upper bound of the range.}

\item{params}{A list containing the mean ('mu') and variance ('sigma') of the Gaussian distribution.}
}
\value{
The probability within the specified range for the Gaussian distribution.
}
\description{
'accelerationProbability' Calculates the probability within a specified acceleration range for a Gaussian acceleration distribution
}
\examples{
# Calculate the probability of observing a value within the acceleration range [1, 3] for a Gaussian distribution with mean = 2 and variance = 1
accelerationProbability(upper = 1, lower = 3, params = list(mu = 2, sigma = 1))

}
