% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMotifProbabilities.R
\name{calculateMotifProbabilities}
\alias{calculateMotifProbabilities}
\title{calculateMotifProbabilities}
\usage{
calculateMotifProbabilities(motifs, filepath)
}
\arguments{
\item{motifs}{A character vector containing the names of motifs for which the motif probability will be calculated. Each name should correspond to an object in the current R environment.}

\item{filepath}{A character string specifying the directory path where the fitted HSMM objects are located. The function will calculate the motif probabilities for all files in this directory.}
}
\value{
A data frame containing the following columns:
\describe{
\item{id}{The first column corresponds to the filename}
\item{motif probabilities}{Columns corresponding to each motif, containing the calculated probabilities.}
}
}
\description{
calculateMotifProbabilities Calculates the probabilities of the specified motifs for each individual using the individually fitted Hidden Semi-Markov Models (HSMMs). For each file in the specified folder, it loads the HSMM and computes the specified motif probabilities.
}
\details{
This function calculates the probabilities of specified motifs for each participant
using fitted Hidden Semi-Markov Models (HSMMs). It loads HSMMs from specified files
and computes motif probabilities for each participant based on their respective model.
}
\examples{
motifs_list <- c("motif1", "motif2")
motif_probabilities <- calculateMotifProbabilities(motifs = motifs_list, filepath = "path/to/hsmm/files")

}
