% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_initialparams.R
\name{derive_initialparams}
\alias{derive_initialparams}
\title{derive_initialparams}
\usage{
derive_initialparams(
  data,
  n_states,
  bout_durations = c(60, 30, 15, 10),
  epoch_length
)
}
\arguments{
\item{data}{A 'hsmm.data' object that includes the accelerometer data.}

\item{n_states}{An integer indicating the number of states of the hsmm.}

\item{bout_durations}{A vector used to derive parameters for the sojourn duration (Poisson) (Default = c(1, 5, 10, 15)).}

\item{epoch_length}{An integer indicating the epoch length in seconds.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{initial_probability}: A vector of length \code{n_states} with the initial probabilities of the semi-Markov chain.
\item \code{transition_matrix}: A transition matrix with the transition probabilities.
\item \code{emission_distribution}: A list indicating the Gaussian observation distribution, consisting of \code{mu} and \code{sigma}.
\item \code{sojourn_distribution}: A list with parameters for the Poisson runlength distribution.
}
}
\description{
'derive_initialparams' derives the initial parameters for the hsmm from the observations
}
\details{
This function calculates initial parameters required for a hidden semi-Markov model (HSMM) based on the provided training data, number of states, bout durations, and epoch length.
}
\examples{
# Define 'data', 'n_states', 'bout_durations', and 'epoch_length'
# result <- derive_initialparams(data, n_states = 3, bout_durations = c(60, 30, 15, 10), epoch_length = 5)

}
