% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_probability.R
\name{duration_probability}
\alias{duration_probability}
\title{duration_probability}
\usage{
duration_probability(duration, lambda)
}
\arguments{
\item{duration}{The duration for which the probability needs to be calculated.}

\item{lambda}{The lambda parameter of the exponential distribution.}
}
\value{
The probability of the given duration in the exponential distribution.
}
\description{
'duration_probability' Calculates the probability of a certain duration in an exponential distribution, with a given duration and lambda parameter
}
\examples{
Calculate the probability of duration 2 in an exponential distribution with lambda = 0.5
duration_probability

}
