% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationProbability.R
\name{durationProbability}
\alias{durationProbability}
\title{durationProbability}
\usage{
durationProbability(duration, shift, lambda)
}
\arguments{
\item{duration}{The duration for which the probability needs to be calculated.}

\item{shift}{The shift parameter of the Poisson distribution.}

\item{lambda}{The lambda parameter of the Poisson distribution.}
}
\value{
The probability of the given duration for the (shifted) Poisson distribution.
}
\description{
\code{durationProbability} Calculates the probability of observing a specific duration for a (shifted) Poisson distribution.
}
\examples{
Calculate the probability of duration of 30 sec (model was trained on 15-sec epoch data) in a Poisson distribution with lambda = 0.5
length = 30/15
#durationProbability(duration = length, shift = 1, lambda = 0.5)

}
