% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emission_probability_within_range.R
\name{emission_probability_within_range}
\alias{emission_probability_within_range}
\title{emission_probability_within_range}
\usage{
emission_probability_within_range(lower_bound, upper_bound, params)
}
\arguments{
\item{lower_bound}{The lower bound of the range.}

\item{upper_bound}{The upper bound of the range.}

\item{params}{A list containing the mean ('mu') and variance ('sigma') of the Gaussian distribution.}
}
\value{
The probability within the specified range for the Gaussian distribution.
}
\description{
'emission_probability_within_range' Calculates the probability within a specified range for a Gaussian emission distribution
}
\examples{
# Calculate the probability within range [1, 3] for a Gaussian distribution with mean = 2 and variance = 1
params <- list(mu = 2, sigma = 1)
emission_probability_within_range(1, 3, params)

}
