% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mhsmm.R
\name{format_mhsmm}
\alias{format_mhsmm}
\title{format_mhsmm}
\usage{
format_mhsmm(data, variable_name)
}
\arguments{
\item{data}{A list containing participant data with elements that include a valid day data.frame object containing metrics.}

\item{variable_name}{A string containing the name of the metric to be selected.}
}
\value{
A 'hsmm.data' object consisting of:
\itemize{
\item \code{df}: A data.frame containing the observed sequence, i.e., all vectors with observations pasted.
\item \code{N}: A vector indicating the length of each observed sequence.
}
}
\description{
'format_mhsmm' formats the valid data list into a hsmm.data object
}
\details{
This function takes a list of participant data and a vector specifying metric names and constructs an 'hsmm.data' object. It extracts the specified metrics from each valid day data frame and concatenates them into a single data frame, also recording the length of each observed sequence.
}
\examples{
# Define 'data' and 'variable_name'
# result_hsmm <- format_mhsmm(data, variable_name = "ENMO")

}
