% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt3x_to_csv.R
\docType{data}
\name{gzip}
\alias{gzip}
\title{gt3x_to_csv}
\format{
An object of class \code{logical} of length 1.
}
\usage{
gzip
}
\arguments{
\item{path}{The directory path where GT3X files are located.}

\item{gzip}{Logical indicating whether to gzip the resulting CSV files (default = TRUE).}
}
\description{
'gt3x_to_csv' converts old ActiGraph GT3X files to CSV format

This script converts ActiGraph GT3X accelerometer files to CSV format using the read.gt3x package.
}
\details{
The script automates the conversion process for all GT3X files in a specified directory. After converting each GT3X file to CSV format, it optionally compresses the CSV files using gzip.   # This code was kindly provided by Hansjorg Baurecht (Regensburg).
}
\keyword{datasets}
