% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_probability.R
\name{motif_probability}
\alias{motif_probability}
\title{motif_probability}
\usage{
motif_probability(motif, hsmm)
}
\arguments{
\item{motif}{A motif is defined as a sequence of activities characterized by their acceleration range and lengths. A data.frame object where the rows represent the activity, with columns Amin, Amax, and length.}

\item{hsmm}{An object containing the Hidden Semi-Markov Model (HSMM) parameters of class 'hsmm'.}
}
\value{
The logarithm of the probability of the entire motif within the hsmm
}
\description{
'motif_probability' Calculates the probability of a motif within a hidden semi-Markov model (hsmm) using the forward algorithm. It requires a motif and the hsmm as inputs.
}
\examples{
# Define 'motif' and 'hsmm' objects
# motif <- data.frame(Amin = c(1, 2, 3), Amax = c(4, 5, 6), length = c(2, 3, 4))
# hsmm <- mhsmm::hsmmfit()
# Calculate motif probability
# motif_probability(motif, hsmm)

}
