% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifProbability.R
\name{motifProbability}
\alias{motifProbability}
\title{motifProbability}
\usage{
motifProbability(motif, hsmm)
}
\arguments{
\item{motif}{A motif is defined as a sequence of bouts, each characterized by their acceleration range and duration. A data.frame object where each rows represent a bout, with columns Amin, Amax, and length.}

\item{hsmm}{An object containing the Hidden Semi-Markov Model (HSMM) parameters of class 'hsmm'.}
}
\value{
The logarithm of the motif probability
}
\description{
'motifProbability' Calculates the probability that a specific sequence of bouts (motif) occurs within the accelerometer data, based on the parameters of the fitted hidden semi-Markov model (hsmm) using the forward algorithm.
}
\examples{
# Define 'motif' and 'hsmm' objects
# motif <- defineMotif(Amin = c(1, 2, 3), Amax = c(4, 5, 6), duration = c(1, 5, 10))
# hsmm <- mhsmm::hsmmfit()
# Calculate motif probability
# motifProbability(motif, hsmm)

}
